backend/pyproject.toml
[tool.poetry]
name = "vnedreid-backend"
version = "0.1.0"
description = ""
authors = ["dorateam"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.16.1"
fastapi = {extras = ["standard"], version = "^0.115.12"}
pyjwt = "^2.10.1"
pydantic-settings = "^2.9.1"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
openai = "^1.84.0"
ruff = "^0.11.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 88
target-version = "py311"
exclude = ["alembic", "__pycache__", "psql_data"]
fix = true
show-fixes = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501", "E711", "E712"]
